pyproject.toml
[tool.poetry]
name = "dewatermark-ai"
version = "0.1.0"
description = ""
authors = ["Kostiakov Aleksandr"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
